requirements.txt
gunicorn
APScheduler==3.10.4
backports.zoneinfo==0.2.1
certifi==2021.5.30
charset-normalizer==2.0.12
click==8.0.4
colorama==0.4.5
cycler==0.11.0
dataclasses==0.8
dnspython==2.2.1
Flask==2.0.3
Flask-APScheduler==1.12.4
Flask-Cors==3.0.10
Flask-WTF==1.0.1
idna==3.4
importlib-metadata==4.8.3
importlib-resources==5.4.0
itsdangerous==2.0.1
Jinja2==3.0.3
kiwisolver==1.3.1
MarkupSafe==2.0.1
matplotlib==3.3.4
numpy==1.19.5
Pillow==8.4.0
pymongo==4.1.1
pyparsing==3.1.0
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2023.4
pytz-deprecation-shim==0.1.0.post0
requests==2.27.1
six==1.16.0
typing_extensions==4.1.1
tzdata==2023.4
tzlocal==4.2
urllib3==1.26.18
Werkzeug==2.0.3
wincertstore==0.2
WTForms==2.3.3
zipp==3.6.0
